refactor(host): split the two Electron files
brain/ipc.ts and ipc/session-acts.ts each held two things: the publication
and act-performing the runtime owns, and the ipcMain registration only an
Electron process can do. Each splits at that line — publication.ts and
session-act-performer.ts on the runtime side, ipc/brain.ts and
ipc/session-acts.ts on the Electron side — so the runtime half carries no
electron import for packages/host to take next. The three words a refused
open answers are declared once and read from both halves.
