+++ apps/desktop/src/main/host/runtime-host.ts
@@ -196,13 +196,13 @@ import { calendarOnboardingOwed } from "../calendar-onboarding-flow";
 import { conversationOperations, startHistoryMaintenance } from "../conversation-operations";
 import { NODE_CAPABILITY } from "../gateway/desktop-node";
 import { createGatewayService, type GatewayService, type GrantedWords } from "../gateway/service";
-import { createSessionActPerformer, NodeAnswerLostError } from "../ipc/session-acts";
 import { wireMemoryMaintenance } from "../memory-maintenance";
 import { type OnboardingState, onboardingStateFile } from "../onboarding-state";
 import { ProviderKeyVaultSync, type VaultSyncAccount } from "../provider-key-vault-sync";
 import type { RunMode } from "../run-mode";
 import { agentRootPath } from "../runtime-store-path";
 import { wireRuntimeStore } from "../runtime-store-wiring";
+import { createSessionActPerformer, NodeAnswerLostError } from "../session-act-performer";
 import { type SecretCipher, SettingsStore } from "../settings-store";
 import { type OnboardingBeatKind, SpeechArbiter } from "../voice/speech-arbiter";
 import { VoiceReceiver } from "../voice-receiver";
